fix: address clang-tidy warnings
bugprone-suspicious-realloc-usage (alloc.h): use a temporary pointer in
REALLOC_ARRAY so the original buffer is not lost if realloc returns NULL.
This was a real memory-leak bug, not just a style issue.

bugprone-suspicious-string-compare (configdep.c, size.c): add explicit
!= 0 to memcmp/strcmp results used as boolean conditions.

clang-analyzer-deadcode.DeadStores (size.c): remove dead `remaining = 0`
store — the outer break already exits the while loop.

clang-analyzer-core.NonNullParamChecker (size.c): guard qsort() calls
with nr > 0 checks so a NULL pointer is never passed.

bugprone-sizeof-expression (size.c): use sizeof(*sorted) instead of
sizeof(sorted[0]) when sorting an array of pointers.

bugprone-implicit-widening-of-multiplication-result (md5.c): cast loop
index to size_t before multiplying to avoid implicit widening.

clang-analyzer-optin.performance.Padding (partition_table.c): reorder
subtype_names struct fields to eliminate padding bytes.

bugprone-branch-clone (partition_table.c): collapse three identical
empty if/else-if branches into a single combined condition.

bugprone-reserved-identifier (posix_exe.c): move -D_POSIX_C_SOURCE to
Makefile POSIX CFLAGS instead of defining the reserved identifier in
source.

+++ md5.c
@@ -46,10 +46,10 @@ static void md5_block(struct md5_ctx *ctx, const uint8_t *block)
 	uint32_t i;
 
 	for (i = 0; i < 16; i++)
-		m[i] = (uint32_t)block[i * 4] |
-		       ((uint32_t)block[i * 4 + 1] << 8) |
-		       ((uint32_t)block[i * 4 + 2] << 16) |
-		       ((uint32_t)block[i * 4 + 3] << 24);
+		m[i] = (uint32_t)block[(size_t)i * 4] |
+		       ((uint32_t)block[(size_t)i * 4 + 1] << 8) |
+		       ((uint32_t)block[(size_t)i * 4 + 2] << 16) |
+		       ((uint32_t)block[(size_t)i * 4 + 3] << 24);
 
 	for (i = 0; i < 64; i++) {
 		uint32_t f, g;
@@ -131,9 +131,9 @@ void md5_final(struct md5_ctx *ctx, uint8_t digest[16])
 	md5_update(ctx, bits, 8);
 
 	for (i = 0; i < 4; i++) {
-		digest[i * 4] = (uint8_t)(ctx->state[i]);
-		digest[i * 4 + 1] = (uint8_t)(ctx->state[i] >> 8);
-		digest[i * 4 + 2] = (uint8_t)(ctx->state[i] >> 16);
-		digest[i * 4 + 3] = (uint8_t)(ctx->state[i] >> 24);
+		digest[(size_t)i * 4] = (uint8_t)(ctx->state[i]);
+		digest[(size_t)i * 4 + 1] = (uint8_t)(ctx->state[i] >> 8);
+		digest[(size_t)i * 4 + 2] = (uint8_t)(ctx->state[i] >> 16);
+		digest[(size_t)i * 4 + 3] = (uint8_t)(ctx->state[i] >> 24);
 	}
 }
